telegram-webhook-bot/dugtrio-webhook-bot/requirements.txt
Flask
python-dotenv
requests
flask-ngrok
telebot